requirements.txt
# Core Framework
streamlit==1.31.0
python-dotenv==1.0.0

# Image Processing
Pillow==10.2.0
opencv-python==4.9.0.80
numpy==1.26.3

# AI/ML Models
google-generativeai==0.3.2
transformers==4.37.2
torch==2.2.0
torchvision==0.17.0

# Data & Utilities
pandas==2.2.0
matplotlib==3.8.2
seaborn==0.13.2
plotly==5.18.0

# Kaggle Integration (for deployment)
pyngrok==7.0.5
kaggle==1.5.16

# Testing
pytest==8.0.0
pytest-cov==4.1.0
